+++ spec/observables/onErrorResumeNext-spec.ts
@@ -1,4 +1,4 @@
-import { onErrorResumeNext } from '../../src/create';
+import { onErrorResumeNext } from '../../src/';
 import marbleTestingSignature = require('../helpers/marble-testing'); // tslint:disable-line:no-require-imports
 
 declare const hot: typeof marbleTestingSignature.hot;
