--- src/create/index.ts
@@ -1,25 +1,0 @@
-/* Static observable creation exports */
-export { bindCallback } from '../internal/observable/bindCallback';
-export { bindNodeCallback } from '../internal/observable/bindNodeCallback';
-export { combineLatest } from '../internal/observable/combineLatest';
-export { concat } from '../internal/observable/concat';
-export { defer } from '../internal/observable/defer';
-export { empty } from '../internal/observable/empty';
-export { forkJoin } from '../internal/observable/forkJoin';
-export { from } from '../internal/observable/from';
-export { fromEvent } from '../internal/observable/fromEvent';
-export { fromEventPattern } from '../internal/observable/fromEventPattern';
-export { generate } from '../internal/observable/generate';
-export { _if as iif } from '../internal/observable/if';
-export { interval } from '../internal/observable/interval';
-export { merge } from '../internal/observable/merge';
-export { never } from '../internal/observable/never';
-export { of } from '../internal/observable/of';
-export { onErrorResumeNext } from '../internal/observable/onErrorResumeNext';
-export { pairs } from '../internal/observable/pairs';
-export { race } from '../internal/observable/race';
-export { range } from '../internal/observable/range';
-export { throwError } from '../internal/observable/throwError';
-export { timer } from '../internal/observable/timer';
-export { using } from '../internal/observable/using';
-export { zip } from '../internal/observable/zip';
