+++ tsconfig/tsconfig.base.json
@@ -5,7 +5,6 @@
     "../dist/src/add/observable/of.ts",
     "../dist/src/index.ts",
     "../dist/src/operators/index.ts",
-    "../dist/src/create/index.ts",
     "../dist/src/ajax/index.ts",
     "../dist/src/websocket/index.ts",
     "../dist/src/testing/index.ts"
